MoviesListPage: refine loading state. Decompose MoviesListPage component.
